requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic-settings==2.4.0
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
dotenv==0.9.9
python-dotenv==1.1.0
pytest==8.3.5
aiohttp==3.11.18
validators==0.35.0
python-magic==0.4.27
pytest-asyncio==0.26.0
httpx==0.28.1
pillow==9.1
numpy==2.2.6
aiobotocore==2.22.0
botocore==1.37.3
aiofiles==24.1.0
openai==1.79.0
alembic==1.15.2
greenlet==3.2.2
asyncpg==0.30.0
# New database and vector store clients
PyMySQL==1.1.0
aiomysql==0.2.0
qdrant-client==1.7.3
PyMuPDF==1.24.9